add a condition to check if the file exist

+++ src/common/file.cleanup-interceptor.ts
@@ -25,16 +25,22 @@ export class FileCleanupInterceptor implements NestInterceptor {
     return next.handle().pipe(
       catchError((error) => {
         if (request.file) {
-          fs.unlinkSync(request.file.path);
+          if (fs.existsSync(request.file.path)) {
+            fs.unlinkSync(request.file.path);
+          }
         } else if (request.files) {
           if (request.files.image) {
             request.files.image.forEach((file) => {
-              fs.unlinkSync(file.path);
+              if (fs.existsSync(file.path)) {
+                fs.unlinkSync(file.path);
+              }
             });
           }
           if (request.files.video) {
             request.files.video.forEach((file) => {
-              fs.unlinkSync(file.path);
+              if (fs.existsSync(file.path)) {
+                fs.unlinkSync(file.path);
+              }
             });
           }
         }
